messages

+++ task_manager/tasks/views.py
@@ -1,5 +1,8 @@
+import pdb
+
 from django.shortcuts import render, HttpResponseRedirect
 from django.views import View
+from django.contrib import messages
 from task_manager.mixins import LoginRequiredWithMessageMixin
 from task_manager.tasks.forms import TaskForm
 from task_manager.tasks.models import Task
@@ -8,9 +11,7 @@ from task_manager.tasks.models import Task
 class TasksView(LoginRequiredWithMessageMixin, View):
     def get(self, request, *args, **kwargs):
         tasks = Task.objects.all()
-        return render(request,
-                      'tasks/index.html',
-                      {'tasks': tasks})
+        return render(request,'tasks/index.html',{'tasks': tasks})
 
 
 class CreateTaskView(LoginRequiredWithMessageMixin, View):
@@ -20,58 +21,50 @@ class CreateTaskView(LoginRequiredWithMessageMixin, View):
                       'tasks/create.html',
                       {'form': form})
 
+
     def post(self, request, *args, **kwargs):
         form = TaskForm(data=request.POST)
         form.instance.user = request.user
         if form.is_valid():
             form.save()
+            messages.info(request, 'Задача успешно создана')
             return HttpResponseRedirect('/tasks')
         else:
-            return render(request,
-                          'tasks/create.html',
-                          {'form': form}
-                          )
+            return render(request,'tasks/create.html',{'form': form})
 
 
 class UpdateTaskView(LoginRequiredWithMessageMixin, View):
     def get(self, request, *args, **kwargs):
-        task = (Task.objects
-                .filter(user_id=request.user.id)
-                .get(id=kwargs['id'])
-                )
+        task = Task.objects.get(id=kwargs['id'])
         form = TaskForm(instance=task)
-        return render(request,
-                      'tasks/update.html',
-                      {'form': form}
-                      )
+        return render(request,'tasks/update.html',{'form': form})
+
 
     def post(self, request, *args, **kwargs):
-        task = (Task.objects
-                .filter(user_id=request.user.id)
-                .get(id=kwargs['id']))
+        task = Task.objects.get(id=kwargs['id'])
         form = TaskForm(instance=task, data=request.POST)
         if form.is_valid():
             form.save()
+            messages.info(request, 'Задача успешно изменена')
             return HttpResponseRedirect('/tasks')
         else:
-            return render(request,
-                          'tasks/update.html',
-                          {'form': form}
-                          )
+            return render(request,'tasks/update.html',{'form': form})
 
 
 class DeleteTaskView(LoginRequiredWithMessageMixin, View):
     def get(self, request, *args, **kwargs):
-        task = (Task.objects
-                .filter(user_id=request.user.id)
-                .get(id=kwargs['id']))
-        return render(request,
-                      'tasks/delete.html',
-                      {'task': task})
+        task = Task.objects.get(id=kwargs['id'])
+        if task.user_id == request.user.id:
+            return render(request,'tasks/delete.html',{'task': task})
+        else:
+            messages.error(request, 'Задачу может удалить только ее автор')
+            return HttpResponseRedirect('/tasks')
+
 
     def post(self, request, *args, **kwargs):
-        task = (Task.objects
-                .filter(user_id=request.user.id)
-                .get(id=kwargs['id']))
-        task.delete()
+        task = Task.objects.get(id=kwargs['id'])
+        if task.user_id == request.user.id:
+            task.delete()
+        else:
+            messages.error(request, 'Задачу может удалить только ее автор')
         return HttpResponseRedirect('/tasks')
